refactor: divide index.ts to app.ts and index.ts, for tests to work - put app.listen to index.ts - remove it from app.ts where is everything else
